Organize e2e tests
